pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asaerc"
version = "0.1.0"
description = "2D PDE reservoir with linear/attention/adaptive-sensing readouts"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name="Felix"}]
dependencies = [
  "numpy>=1.24",
  "torch>=2.1",
  "tqdm>=4.66",
  "matplotlib>=3.7",
  "joblib>=1.3",
  "pyyaml>=6.0",
  "scipy>=1.10",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.6",
  "black>=24.0",
  "mypy>=1.8",
  "pre-commit>=3.6",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
testpaths = ["tests"]
